Fix sign-unsigned comparison.
This is flagged by `-Wsign-compare` which we do not typically activate.

(cherry picked from commit 28acc037ea7d2a48f22c106e73d62465eec8ab5f)

+++ hilti/runtime/include/types/stream.h
@@ -1285,7 +1285,7 @@ public:
             throw WouldBlock("end of stream view");
 
         auto p = unsafeBegin();
-        for ( int i = 0; i < n; ++i )
+        for ( uint64_t i = 0; i < n; ++i )
             dst[i] = *p++;
 
         return View(SafeConstIterator(p), _end);
